Fix FileReader bug: one long+multiple short lines
Example:

}abcdefghijklmnopqrstuvw
\n
\n
\n
\n
\n

Previously, the next lines were implemented using a LimitedQueue as
well. This lead to positioning inconsistencies when the maxLookAhead
size of the LimitedQueue was not sufficient. Take the example and
imagine we are at the start with our normal position. The HEAD is some
lines below. advanceHead() will read a new line and thus replace the
first line at some point although we are still in that line with our
"normal" position.

Fixed by using an ArrayList of Strings for the next lines that might
have more entries than the LimitedQueue of the next chars.