Implement & Test Java Lexer (#24)
* Bump maven to Java 15

* Implement Lexer

* Adjust MaxLookAheadException

* Add getters for pos line and col in FileReader

* Rename InvalidSectionException to UnknownCharException

* Add spaces in JavaDoc

* Increase max look ahead by one

* Use String instead of Character[] for lookAheads()

Also changed the mechanism for the TestFileReader used to read in files
in test cases.

* Init LexerTest

Added equals(), hashCode() and toString() to Token

* Explicitly convert EOF to char in FileReaderTest

* Test IF token and fix Lexer bugs

- Set context at the beginning to section if we have a section at the
beginning
- Check EOF after having skipped whitespaces
- in consumeSection(): Check for Tokens first that require more than
one character look ahead
- Add missing() reader.advance() in while loop
(consumeIdentifierOrKeyword()

* Improve test for IF token and add test for HAS

* Add test for NOT token

* Test tokens TRUE and FALSE

* Test operator Tokens, e.g. EQUALS

* Ignore payload comment identifier

* Adjust Token position to be the start not the end.

Until now, the Token position denoted the position in the code where
the Token ended. Changed this, so that the position is the beginning.

Adjusted LexerTests accordingly.

* Test Token IDENTIFIER

* Replace callbacks with member variable

Used to store the Token start position.

* Add JavaDocs to Token start position variables

* Improve token start position handling

* Test Token NUMBER & STRING

* Test Token DOT, BRACES and INDEX

* Fix FileReader bug: one long+multiple short lines

Example:

}abcdefghijklmnopqrstuvw
\n
\n
\n
\n
\n

Previously, the next lines were implemented using a LimitedQueue as
well. This lead to positioning inconsistencies when the maxLookAhead
size of the LimitedQueue was not sufficient. Take the example and
imagine we are at the start with our normal position. The HEAD is some
lines below. advanceHead() will read a new line and thus replace the
first line at some point although we are still in that line with our
"normal" position.

Fixed by using an ArrayList of Strings for the next lines that might
have more entries than the LimitedQueue of the next chars.

* Add advanced HTML Test & limit skip whitespace

Only skip whitespace if we are not in the ARBITRARY context.

* Add advanced Java Lexer test

* Rename Java Test to reflect class name

* Replace String concatenation with text blocks

* Rename String test case (due to name conflict)

* Fix typo "input"

* Simplify consumeBraceClose()

* Add example to ARBITRARY TokenType

* Add arbitrary string between CCom sections (test)

(when using comment line identifiers)
Also adjusted the advanced Java test accordingly.

* Call consumeCommentPayloadIdentifier()

* Use local variable instead of look ahead

* Add code position to UnknownCharException

* Get rid of unnecessary context change

* Inline constructToken return

* Clarify single statements comment

* Create constant EOF for (char) -1

* Implement expectMultiple() in FileReader

* Rename FileReader to Reader

* Add more JavaDocs to Lexer, Token and TokenType

* Test EOF in basic file

* Test expectMultiple() (FileReader)

* Change occurrences of FileReader to Reader

* Test expect multiple failure